requirements.txt
# Data Handling
pandas==2.2.1
numpy==1.26.4
ta==0.10.2

# Dataset Loading
kagglehub==0.1.6
kaggle==1.6.6
yfinance
requests>=2.31
multitasking>=0.0.7
platformdirs>=2.0.0
pytz>=2022.5
frozendict>=2.3.4
beautifulsoup4>=4.11.1
peewee>=3.16.2
requests_cache>=1.0
requests_ratelimiter>=0.3.1
scipy>=1.6.3

# Image/Chart Processing
opencv-python==4.9.0.80
matplotlib==3.8.4
tensorflow==2.19.0
torch==2.2.2
torchvision==0.17.2

# LLaMA + RAG Stack(still under exploration to turn into multi-modal AI system)
llama-cpp-python==0.2.38
transformers==4.39.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0  # For vector similarity search
langchain==0.1.14
langchain-openai>=0.0.5
chromadb==0.4.24
tiktoken==0.5.2
langchainhub==0.1.15
openai==1.14.3 
tqdm==4.66.2
langgraph==0.4.1

# Environment
python-dotenv==1.0.1  # For managing .env API keys